requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-mcp-adapters>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
duckdb>=1.0.0
pandas>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
